requirements.txt
Flask==2.3.4
Flask-Cors==3.0.10
python-dotenv==1.1.1
google-genai==0.4.0
requests==2.31.0